Add files via upload
COMPUTER GRAPHICS MINI PROJECT :
Main aim of this Mini Project is to illustrate the concepts and usage of pre-built functions in OpenGL.  Simulation approaches oriented  to the evaluation of complex configurations of bus stops used as evaluation tools for the design and planning of  bus lines  are  also  described. It allows the user to start the bus and picking up the passenger, later, the bus drops off at the location and continues to move.  It uses GLUT Library and OpenGL functions for proper function of the code.